backend/requirements.txt
black==23.9.1
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
mixer==7.2.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
